feat: Refactors rounds state
The rounds state was not setup well to help that historic, current and
future shapes. While this means that a bad state may exist where there
are more than 5 rounds, it prioritises having the current round state
pristine and without any unnecessary partiality when processing an event
in the rounds state.